requirements.txt
discord.py~=2.3.0
firebase~=4.0.0
